+++ DaysSinceShared/DaysSinceShared.h
@@ -19,7 +19,7 @@ FOUNDATION_EXPORT const unsigned char DaysSinceSharedVersionString[];
 
 NSURL *applicationDocumentsDirectory();
 
-NSString *pathToStoredCounter(int counterNum);
+NSString *pathToStoredCounter(NSInteger counterNum);
 
 NSString *pathToCurrentViewIndex();
 
@@ -27,4 +27,4 @@ NSString *pathToCurrentViewIndex();
  * 
  * Should only be called on application startup, at which point the value should be stored and maintained by the view controller.
  */
-int numStoredCounters();
+NSInteger numStoredCounters();
